requirements.txt
aedat==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
bbox-visualizer==0.1.0
build==1.3.0
cachetools==5.5.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
docker-pycreds==0.4.0
einops==0.8.0
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
h5py==3.11.0
human-body-prior==2.2.2.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.7
imageio==2.34.1
importlib_metadata==8.7.1
importlib_resources==6.4.0
iopath==0.1.10
Jinja2==3.1.3
kiwisolver==1.4.5
lightning-utilities==0.11.2
loguru==0.7.2
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.3
nvidia-ml-py==12.560.30
nvitop==1.4.0
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.18.0
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.2.0
pip-tools==7.5.2
plotly==5.22.0
portalocker==2.8.2
protobuf==4.25.3
psutil==5.9.8
pycocotools==2.0.7
pyparsing==3.1.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytorch-lightning==1.8.6
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@bd52f4a408b29dc6b4357b70c93fd7a9749ca820
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
resample2d-cuda==0.0.0
scipy==1.13.0
seaborn==0.13.2
sentry-sdk==2.0.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
StrEnum==0.4.15
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboardX==2.6.2.2
termcolor==2.4.0
tomli==2.3.0
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchdata==0.7.1
torchmetrics==1.3.2
torchvision==0.16.0+cu118
tqdm==4.66.4
triton==2.1.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.6
yacs==0.1.8
yarl==1.9.4
zipp==3.23.0
